+++ src/main/java/umc/cockple/demo/domain/chat/handler/ChatWebSocketHandler.java
@@ -16,7 +16,7 @@ import umc.cockple.demo.domain.chat.events.ChatMessageSendEvent;
 import umc.cockple.demo.domain.chat.events.ChatRoomSubscriptionEvent;
 import umc.cockple.demo.domain.chat.exception.ChatException;
 import umc.cockple.demo.domain.chat.service.ChatValidator;
-import umc.cockple.demo.domain.chat.service.SubscriptionService;
+import umc.cockple.demo.domain.chat.service.websocket.SubscriptionService;
 import umc.cockple.demo.domain.member.service.MemberQueryService;
 
 import java.time.LocalDateTime;
